View: Verbesserungen, images hinzugefügt, Border, Fullscreen

+++ KneeDeepInTheBlood/src/model/Position.java
@@ -7,21 +7,21 @@ package model;
  */
 public class Position 
 {
-	private int x = 0;
-	private int y = 0;
-	public int getX() {
+	private double x = 0;
+	private double y = 0;
+	public double getX() {
 		return x;
 	}
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
-	public Position(int x, int y) {
+	public Position(double x, double y) {
 		super();
 		this.x = x;
 		this.y = y;
